+++ src/aov1.jl
@@ -10,10 +10,10 @@ function aov1(x, Y)
     q = size(Y, 2) 
     A = length(unique(x))
     Xdummy = dummy(x).Y
-    zY = center(Y, colmeans(Y))
+    zY = center(Y, colmean(Y))
     fm = mlr(Xdummy, zY) ;
     pred = predict(fm, Xdummy).pred
-    SSF = sum((pred.^2), dims = 1)   # = colvars(pred) * n
+    SSF = sum((pred.^2), dims = 1)   # = colvar(pred) * n
     SSR = ssr(pred, zY)
     df_fact = A - 1 
     df_res = n - A